fix some C4244 warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4724 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/DSMSplitter/DSMSplitterFile.cpp
@@ -241,7 +241,7 @@ bool CDSMSplitterFile::Read(__int64 len, IDSMResourceBagImpl& res)
 		return false;    // TODO
 	}
 
-	r.data.SetCount(len);
+	r.data.SetCount((size_t)len);
 	ByteRead(r.data.GetData(), r.data.GetCount());
 
 	res += r;
